refactor: ReservationDao 구조 단순화
Mapper 및 ResultMap 제거
