+++ src/main/java/roomescape/dao/ReservationDao.java
@@ -1,17 +1,17 @@
-package roomescape.dao.reservation;
+package roomescape.dao;
 
 import java.util.List;
 import java.util.Objects;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.EmptyResultDataAccessException;
-import org.springframework.jdbc.core.DataClassRowMapper;
 import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
+import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
+import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 
 import roomescape.domain.Reservation;
+import roomescape.domain.Time;
 import roomescape.exception.IdNotExistException;
 import roomescape.exception.TimeNotExistException;
 
@@ -31,27 +31,26 @@ public class ReservationDao {
 
     public List<Reservation> getAllReservations() {
         return jdbcTemplate.query("""
-                    SELECT 
-                    r.id as reservation_id, 
-                    r.name, 
-                    r.date, 
-                    t.id as time_id, 
-                    t.time as time_value 
-                FROM reservation as r inner join time as t on r.time_id = t.id
-                """, new DataClassRowMapper<>(ReservationResultMap.class))
-            .stream()
-            .map(ReservationResultMapMapper::toDomain)
-            .toList();
+                SELECT 
+                r.id as reservation_id, 
+                r.name, 
+                r.date, 
+                t.id as time_id, 
+                t.time as time_value 
+            FROM reservation as r inner join time as t on r.time_id = t.id
+            """, ((rs, count) -> new Reservation(
+            rs.getLong("reservation_id"),
+            rs.getString("name"),
+            rs.getString("date"),
+            new Time(rs.getLong("time_id"), rs.getString("time_value"))
+        )));
     }
 
     public Long saveReservation(Reservation reservation) {
         validateSaveReservation(reservation);
         String time = getTime(reservation);
-
-        var parameterSource = new BeanPropertySqlParameterSource(
-            ReservationResultMapMapper.toResultMap(reservation));
+        SqlParameterSource parameterSource = getReservationParameterSource(reservation);
         Number id = simpleJdbcInsert.executeAndReturnKey(parameterSource);
-
         reservation.setId(id.longValue());
         reservation.setTimeValue(time);
         return id.longValue();
@@ -88,5 +87,14 @@ public class ReservationDao {
             jdbcTemplate.queryForObject("SELECT EXISTS (SELECT 1 FROM reservation WHERE id = " + id + ")",
                 Boolean.class)
         );
+    }
+
+    private MapSqlParameterSource getReservationParameterSource(Reservation reservation) {
+        return new MapSqlParameterSource()
+            .addValue("reservationId", reservation.getId())
+            .addValue("name", reservation.getName())
+            .addValue("date", reservation.getDate())
+            .addValue("timeId", reservation.getTimeId())
+            .addValue("timeValue", reservation.getTimeValue());
     }
 }
