--- src/main/java/roomescape/dao/reservation/ReservationResultMap.java
@@ -1,4 +1,0 @@
-package roomescape.dao.reservation;
-
-public record ReservationResultMap(Long reservationId, String name, String date, Long timeId, String timeValue) {
-}
